requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bleach==6.1.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.48.1
gast==0.4.0
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
kaggle==1.6.5
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opendatasets==0.1.22
opt-einsum==3.3.0
pandas==2.2.0
pillow==10.2.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-slugify==8.0.3
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
tqdm==4.66.1
tzdata==2023.4
urllib3==2.2.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0

